+++ jobquest/model.py
@@ -2,7 +2,7 @@ from __future__ import annotations
 
 from abc import ABC, abstractmethod
 from typing import Optional
-from datetime import date
+from datetime import date, datetime
 from enum import Enum
 from pydantic import BaseModel
 
@@ -246,7 +246,7 @@ class JobPost(BaseModel):
 
     job_type: list[JobType] | None = None
     compensation: Compensation | None = None
-    date_posted: date | None = None
+    date_posted: datetime | date | None = None
     emails: list[str] | None = None
     is_remote: bool | None = None
     listing_type: str | None = None